Small refactor to move monarchs directory into `src` directory, and tidied up .gitignore and .dockerignore
