+++ .dockerignore
@@ -2,13 +2,13 @@ __pycache__/
 .idea/
 .run
 progress/
-monarchs/plots/outputs/
+src/monarchs/plots/outputs/
 venv/
-monarchs/data/RACMO_temp.nc
-monarchs/data/ERA5_data_subset.nc
+src/monarchs/data/RACMO_temp.nc
+src/monarchs/data/ERA5_data_subset.nc
 data/*.tif
 data/*.jpg
-!monarchs/data/ERA5_small.nc
+!src/monarchs/data/ERA5_small.nc
 *.gif
 *.jpg
 *.m
