+++ .gitignore
@@ -2,12 +2,12 @@ __pycache__/
 .idea/
 .run
 progress/
-monarchs/plots/outputs/
+src/monarchs/plots/outputs/
 venv/
-monarchs/DEM/38_12_32m_v2.0
-monarchs/data/RACMO_temp.nc
-monarchs/data/ERA5_data_subset.nc
-monarchs/data/38_12_32m_v2.0_meta.txt
+src/monarchs/DEM/38_12_32m_v2.0
+src/monarchs/data/RACMO_temp.nc
+src/monarchs/data/ERA5_data_subset.nc
+src/monarchs/data/38_12_32m_v2.0_meta.txt
 data/*.tif
 data/*.jpg
 plots/
