+++ run_MONARCHS.py
@@ -1,4 +1,4 @@
-from monarchs.core.driver import monarchs
+from src.monarchs.core.driver import monarchs
 
 if __name__ == "__main__":
     grid = monarchs()
