+++ tests/serial/test_firn_column.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from monarchs.met_data.import_ERA5 import (
+from src.monarchs.met_data.import_ERA5 import (
     ERA5_to_variables,
     grid_subset,
     interpolate_grid,
